move config to service bus app.  make config accessible via shell.  merge config in config file.
