+++ deploy/service-bus/delete.js
@@ -16,7 +16,7 @@ try {
 }
 const azure = require('azure')
 const sbService = azure.createServiceBusService(process.env.AZURE_SERVICE_BUS_CONNECTION_STRING)
-const queues = require('../deploy.config').ServiceBus.Queues
+const queues = require('./deploy.config')
 
 const deleteQueue = (queueName) => (new Promise((resolve, reject) => {
   // There is no counterpart to createQueueIfNotExists, although deleteQueueIfNotExists is available
